+++ src/GitHub_Updater/Settings.php
@@ -385,15 +385,7 @@ class Settings extends Base {
 			$type                             = '<span class="dashicons dashicons-admin-plugins"></span>&nbsp;';
 			$setting_field                    = array();
 			$ghu_options_keys[ $token->repo ] = null;
-
-			/*
-			 * Set defaults if repo_meta not set and no update available.
-			 */
-			if ( empty( $token->repo_meta ) ) {
-				$defaults = $this->set_defaults( $token->type );
-				unset( $defaults['repo'] );
-				$token = (object) array_merge( (array) $token, $defaults );
-			}
+			$token->private                   = isset( $token->private ) ? $token->private : true;
 
 			/*
 			 * Set boolean for Enterprise headers.