+++ CHANGES.md
@@ -4,7 +4,7 @@
 * switched to modified version of [wp.org plugin readme parser](https://meta.trac.wordpress.org/browser/sites/trunk/wordpress.org/public_html/wp-content/plugins/plugin-directory/readme/class-parser.php), now accepts _Markdownified_ readme.txt files
 * fixed re-activation of RESTful plugin update, multisite vs single site
 * when creating Settings page, check current Plugin/Theme class instance, not transient. Fixes issue where remote install of private repo not having private settings saved.
-* add default repo values in Settings page to fix PHP errors
+* fixed PHP errors in Settings page
 
 #### 5.5.0
 * better internationalization for changing plugin _View details_ link